refactor(war): split SipCallHandler into focused single-responsibility beans
SipCallHandler (700 lines) was doing too much: it owned per-call
state, played the selection menu, ran the announcement loop, handled
DTMF from two sources, and managed call teardown on BYE and shutdown.

Split into six focused classes:

CallState           — immutable per-call data record (was a private
                      inner record; now package-private file)
SipCallHeaders      — static utilities: callPrefix, normaliseHeader,
                      buildCallerIdentitySummary
CallSessionManager  — session store (activeCalls, pendingSelections),
                      BYE handling, Liberty @PreDestroy shutdown
AnnouncementLoop    — time-announcement playback loop; exits on
                      interrupt, socket close, or CALL_MAX_DURATION
MenuRunner          — language-selection menu loop; transitions to
                      AnnouncementLoop on digit or hangs up on timeout
DtmfDispatcher      — DTMF handling for both SIP INFO (RFC 2976) and
                      RFC 2833 telephone-event RTP; starts the RFC 2833
                      receiver thread
CallAcceptor        — INVITE acceptance, blacklist check, language menu
                      building, SDP negotiation

SipCallHandler itself is now a thin 50-line delegate that only routes
INVITE → CallAcceptor, INFO → DtmfDispatcher, BYE → CallSessionManager.
The SipTimeServlet entry point is unchanged.

Also cleaned up a dead inner-check (Instant.now().isAfter(deadline)
inside a while that already guards on the same condition) and renamed
catch (Exception ignored) to catch (Exception ignoredException).
